+++ packages/zowe-explorer/__tests__/__unit__/job/init.unit.test.ts
@@ -52,6 +52,7 @@ describe("Test src/jobs/extension", () => {
             ssoLogin: jest.fn(),
             ssoLogout: jest.fn(),
             onDidChangeConfiguration: jest.fn(),
+            onDidCloseTextDocument: jest.fn(),
             pollData: jest.fn(),
             refreshElement: jest.fn(),
             filterJobsDialog: jest.fn(),
@@ -59,8 +60,8 @@ describe("Test src/jobs/extension", () => {
         const commands: IJestIt[] = [
             {
                 name: "zowe.jobs.zosJobsOpenspool",
-                parm: [test._, test.value, test._],
-                mock: [{ spy: jest.spyOn(jobActions, "getSpoolContent"), arg: [test._, test.value, test._] }],
+                parm: [test._, test.value],
+                mock: [{ spy: jest.spyOn(jobActions, "getSpoolContent"), arg: [test._, test.value] }],
             },
             {
                 name: "zowe.jobs.deleteJob",
@@ -242,6 +243,7 @@ describe("Test src/jobs/extension", () => {
 
             spyCreateJobsTree.mockResolvedValue(jobsProvider as any);
             spyOnSubscriptions(commands);
+            jest.spyOn(vscode.workspace, "onDidCloseTextDocument").mockImplementation(jobsProvider.onDidCloseTextDocument);
             await initJobsProvider(test.context);
         });
         beforeEach(() => {
@@ -257,6 +259,10 @@ describe("Test src/jobs/extension", () => {
             spyCreateJobsTree.mockResolvedValue(null);
             const myProvider = await initJobsProvider({} as any);
             expect(myProvider).toBe(null);
+        });
+
+        it("should register onDidCloseTextDocument event listener from ZosJobsProvider", () => {
+            expect(jobsProvider.onDidCloseTextDocument).toHaveBeenCalledWith(jobTree.ZosJobsProvider.onDidCloseTextDocument);
         });
     });
 });