+++ packages/zowe-explorer/src/job/ZoweJobNode.ts
@@ -179,7 +179,7 @@ export class Job extends ZoweTreeNode implements IZoweJobTreeNode {
                     spoolNode.command = {
                         command: "zowe.jobs.zosJobsOpenspool",
                         title: "",
-                        arguments: [sessionName, spool, refreshTimestamp],
+                        arguments: [sessionName, spoolNode],
                     };
                     elementChildren[newLabel] = spoolNode;
                 }