+++ packages/zowe-explorer/src/uss/init.ts
@@ -196,6 +196,7 @@ export async function initUSSProvider(context: vscode.ExtensionContext): Promise
             await ussFileProvider.onDidChangeConfiguration(e);
         })
     );
+    context.subscriptions.push(vscode.workspace.onDidCloseTextDocument(USSTree.onDidCloseTextDocument));
 
     initSubscribers(context, ussFileProvider);
     return ussFileProvider;