+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 - Added new optional boolean parameter `hideFromAllTrees` to `IZoweTree.deleteSession` for specifying whether to hide from all trees or current tree. [#2567](https://github.com/zowe/vscode-extension-for-zowe/issues/2567)
 - Added new optional parameter `provider` of type `IZoweTree<IZoweTreeNode>` for `IZoweTree.addSession` to specify a tree to add the profile to.
 - Added optional `filter` and `actualJobs` variables to `IZoweJobTreeNode` to track local filter search.
+- Added new optional record `openFiles` to `IZoweTree` to track opened files under a specific tree view. [#2597](https://github.com/zowe/vscode-extension-for-zowe/issues/2597)
 
 ### Bug fixes